Split graphics logic
There are still some graphics parts in the menu logic and in solid
masks, but those will be moved later.
